refactor: update filenames and docstrings
